Migrate project handlers into projectevents.js
