+++ src/handler.js
@@ -3,6 +3,7 @@ import Render from './render';
 import Storage from './storage';
 import Project from './project';
 import FormEvents from './formevents';
+import ProjectEvents from './projectevents';
 
 export default class Handler {
   // TASK EVENT HANDLERS CODE BLOCKS BELOW. //
@@ -35,108 +36,11 @@ export default class Handler {
     task.remove();
   }
 
-  // PROJECT EVENT HANDLERS CODE BLOCK BELOW. //
-  static projectEvents() {
-    Handler.addProjectButtons();
-  }
-
-  static addProjectButtons() {
-    const buttonAddProject = document.querySelector('#button-add-project');
-    const buttonCancelProject = document.querySelector('#button-cancel-project');
-
-    buttonAddProject.addEventListener('click', Handler.addProject);
-    buttonCancelProject.addEventListener('click', Handler.cancelProject);
-  }
-
-  static addProject() {
-    const projectName = document.querySelector('#input-add-project').value;
-
-    if (projectName === '') {
-      console.log('Project name cannot be empty!');
-      return;
-    }
-    if (Storage.getProjectsList().contains(projectName)) {
-      console.log(`${projectName} already exist!`);
-      return;
-    }
-
-    Storage.addProject(new Project(projectName));
-    Render.aProjectCard(projectName);
-    Handler.initProjectButtons();
-    // console.log(projectName);
-  }
-
-  static cancelProject() {
-    const inputAddProject = document.querySelector('#input-add-project');
-    inputAddProject.value = '';
-    // console.log('project cancelled');
-  }
-
-  // TODO: when pressed, opens project in main container
-  static selectedProjectInSidebar(e) {
-
-  }
-
-  static initProjectButtons() {
-    // const inboxProjectsButton = document.querySelector('#button-inbox-projects');
-    // inboxProjectsButton.addEventListener('click', Handler.openInboxTasks);
-
-    const buttonProjects = document.querySelectorAll('.button-project');
-    buttonProjects.forEach((buttonProject) => {
-      buttonProject.addEventListener('click', Handler.selectedProjectInSidebar);
-    });
+  static project() {
+    ProjectEvents.projectEvents();
   }
 
   static form() {
     FormEvents.formEvents();
   }
-  // // FORM EVENT HANDLERS CODE BLOCKS BELOW. //
-
-  // static formEvents() {
-  //   Handler.forAddTaskButton();
-  //   Handler.forTaskFormButtons();
-  // }
-
-  // static forAddTaskButton() {
-  //   const addTaskButton = document.querySelector('#add-task');
-  //   addTaskButton.addEventListener('click', Handler.openTaskForm);
-  // }
-
-  // static openTaskForm() {
-  //   const overlay = document.querySelector('#overlay');
-  //   const formContainer = document.querySelector('#form-container');
-
-  //   overlay.classList.add('active');
-  //   formContainer.classList.add('active');
-  // }
-
-  // static forTaskFormButtons() {
-  //   // const createTaskButton = document.querySelector('#create-task');
-  //   const cancelTaskButton = document.querySelector('#cancel-task');
-  //   const form = document.querySelector('#form');
-
-  //   cancelTaskButton.addEventListener('click', Handler.cancelTaskForm);
-  //   form.addEventListener('submit', Handler.createTask);
-  // }
-
-  // static cancelTaskForm() {
-  //   const overlay = document.querySelector('#overlay');
-  //   const formContainer = document.querySelector('#form-container');
-
-  //   overlay.classList.remove('active');
-  //   formContainer.classList.remove('active');
-  // }
-
-  // static createTask(event) {
-  //   event.preventDefault();
-  //   // const projectName = document.querySelector('#project-name').textContent;
-  //   const taskName = document.querySelector('#task-title').value;
-  //   const date = document.querySelector('#task-date').value;
-
-  //   const newTask = new Task(taskName, date);
-  //   Render.aTaskCard(newTask.getName(), newTask.getDate());
-
-  //   document.querySelector('#form-container').classList.remove('active');
-  //   document.querySelector('#overlay').classList.remove('active');
-  // }
 }
